requirements.txt
asgiref==3.9.1
Django==5.2.4
sqlparse==0.5.3
django-cors-headers==4.7.0
django-filter==25.1
django-rest-framework==0.1.0
djangorestframework==3.16.0
pillow==11.3.0
pip==24.0
sqlparse==0.5.3